Added a spider
Added in spider enemies with animations for idle, walking, attacking and death.
I have modified some of the code to allow for this to work.
Modified the code so when the player gets out of range of the enemy it will stop moving
The enemy will not chase the player even if it is classed as "dead"
Added a skybox that has movement